# src/config/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Banking System API",
    "description": "Enterprise-grade REST API for a banking system with ACID transactions, double-entry ledger, AI-powered fraud detection (Gemini), and admin management. Built with Express 5, MongoDB, and Zod validation.",
    "version": "1.0.0",
    "contact": {
      "name": "Jeel Delvadiya"
    }
  },
  "servers": [
    { "url": "http://localhost:3000", "description": "Development server" }
  ],
...
  "paths": {
...
    "/api/accounts": {
      "get": {
        "tags": ["Accounts"],
        "summary": "List all accounts of the authenticated user",
        "security": [{ "cookieAuth": [] }],
        "responses": {
          "200": {
            "description": "Array of user's accounts with balances",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "accounts": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Account" }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Accounts"],
        "summary": "Create a new bank account",
        "security": [{ "cookieAuth": [] }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "currency": { "type": "string", "enum": ["INR", "USD", "EUR"], "default": "INR" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Account created successfully" }
        }
      }
    },
...
    "/api/transactions": {
      "get": {
        "tags": ["Transactions"],
        "summary": "Get transaction history (paginated + filterable)",
        "security": [{ "cookieAuth": [] }],
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 10 } },
          { "name": "type", "in": "query", "schema": { "type": "string", "enum": ["sent", "received"] }, "description": "Filter by sent or received" },
          { "name": "status", "in": "query", "schema": { "type": "string", "enum": ["pending", "completed", "failed", "reversed"] } },
          { "name": "fromDate", "in": "query", "schema": { "type": "string", "format": "date" } },
          { "name": "toDate", "in": "query", "schema": { "type": "string", "format": "date" } },
          { "name": "minAmount", "in": "query", "schema": { "type": "number" } },
          { "name": "maxAmount", "in": "query", "schema": { "type": "number" } }
        ],
        "responses": {
          "200": {
            "description": "Paginated transaction list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "transactions": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Transaction" }
                    },
                    "pagination": { "$ref": "#/components/schemas/Pagination" }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Transactions"],
        "summary": "Create a new transfer (with AI fraud detection)",
        "description": "Transfers funds between two accounts. Each transaction passes through a 7-rule fraud detection engine + Gemini AI analysis before execution. Uses MongoDB ACID sessions and double-entry ledger for data integrity.",
        "security": [{ "cookieAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["fromAccount", "toAccount", "amount", "idempotencyKey"],
                "properties": {
                  "fromAccount": { "type": "string", "description": "Source account ID" },
                  "toAccount": { "type": "string", "description": "Destination account ID" },
                  "amount": { "type": "number", "minimum": 1, "example": 5000 },
                  "idempotencyKey": { "type": "string", "format": "uuid", "description": "Unique key to prevent duplicate transactions" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Transfer completed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "message": { "type": "string" },
                    "transaction": { "$ref": "#/components/schemas/Transaction" }
                  }
                }
              }
            }
          },
          "400": { "description": "Insufficient balance, inactive account, or validation error" },
          "429": { "$ref": "#/components/responses/RateLimited" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "token",
        "description": "JWT token set as httpOnly cookie on login"
      }
    },
    "schemas": {
...
      "Account": {
        "type": "object",
        "properties": {
          "_id": { "type": "string" },
          "user": { "type": "string" },
          "currency": { "type": "string", "enum": ["INR", "USD", "EUR"] },
          "status": { "type": "string", "enum": ["active", "frozen", "closed"] },
          "balance": { "type": "number" },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
      "Transaction": {
        "type": "object",
        "properties": {
          "_id": { "type": "string" },
          "fromAccount": { "type": "string" },
          "toAccount": { "type": "string" },
          "amount": { "type": "number" },
          "status": { "type": "string", "enum": ["pending", "completed", "failed", "reversed"] },
          "riskScore": { "type": "number", "description": "AI-calculated fraud risk score (0-100)" },
          "fraudAlert": { "type": "string", "description": "Reference to fraud alert if flagged" },
          "idempotencyKey": { "type": "string", "format": "uuid" },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
      "Pagination": {
        "type": "object",
        "properties": {
          "page": { "type": "integer" },
          "limit": { "type": "integer" },
          "total": { "type": "integer" },
          "totalPages": { "type": "integer" }
        }
      }
    },
    "responses": {
...
      "RateLimited": {
        "description": "Too many requests — rate limit exceeded"
      }
    }
  }
}